feat: add translation manager and symbol highlighter
Implement TranslationsManager to manage JSON translations and display them in a tree view. Integrate JsonSymbolProvider for symbol management. Add highlightSymbolInDocumentAsync to highlight selected symbols in JSON documents, enhancing navigation and management of translation files.
