+++ src/libs/feature/feature-json-symbol-provider/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/json-symbol-provider';
