+++ src/libs/feature/feature-json-tree-data-provider/src/index.ts
@@ -1,0 +1,2 @@
+export * from './lib/json-tree-data-provider';
+export * from './lib/json-tree-item';
