+++ src/libs/feature/feature-translations-manager/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/translations-manager';
