+++ src/libs/util/util-highlight-symbol-in-document/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/highlight-symbol-in-document';
