+++ src/libs/util/util-open-file-by-uri/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/open-file-by-uri';
