+++ src/libs/util/util-open-file-by-uri/src/lib/open-file-by-uri.ts
@@ -1,0 +1,9 @@
+import vscode from 'vscode';
+
+export async function openFileByUriAsync(
+  uri: vscode.Uri
+): Promise<vscode.TextDocument> {
+  const document = await vscode.workspace.openTextDocument(uri);
+  await vscode.window.showTextDocument(document);
+  return document;
+}
