+++ tsconfig.json
@@ -58,8 +58,17 @@
       "@i18n-weave/feature/feature-i18next-scanner-service": [
         "src/libs/feature/feature-i18next-scanner-service/src"
       ],
+      "@i18n-weave/feature/feature-json-symbol-provider": [
+        "src/libs/feature/feature-json-symbol-provider/src"
+      ],
+      "@i18n-weave/feature/feature-json-tree-data-provider": [
+        "src/libs/feature/feature-json-tree-data-provider/src"
+      ],
       "@i18n-weave/feature/feature-translation-service": [
         "src/libs/feature/feature-translation-service/src"
+      ],
+      "@i18n-weave/feature/feature-translations-manager": [
+        "src/libs/feature/feature-translations-manager/src"
       ],
       "@i18n-weave/feature/feature-module-chain-manager": [
         "src/libs/feature/feature-module-chain-manager/src"
@@ -77,6 +86,12 @@
       ],
       "@i18n-weave/util/util-localization-utilities": [
         "src/libs/util/util-localization-utilities/src"
+      ],
+      "@i18n-weave/util/util-highlight-symbol-in-document": [
+        "src/libs/util/util-highlight-symbol-in-document/src"
+      ],
+      "@i18n-weave/util/util-open-file-by-uri": [
+        "src/libs/util/util-open-file-by-uri/src"
       ],
       "@i18n-weave/util/util-array-utilities": [
         "src/libs/util/util-array-utilities/src"
